Changed eventId to int type to be able to use sqlite

+++ EventManager.Server/ApiModels/EventController/EventDto.cs
@@ -2,6 +2,6 @@
 {
     public class EventDto : EventCreateDto
     {
-        public ulong EventId { get; set; }
+        public int EventId { get; set; }
     }
 }
